+++ aioxmpp/bookmarks/__init__.py
@@ -19,7 +19,8 @@
 # <http://www.gnu.org/licenses/>.
 #
 ########################################################################
-""":mod:`~aioxmpp.bookmarks` – Bookmark support (:xep:`0048`)
+"""
+:mod:`~aioxmpp.bookmarks` – Bookmark support (:xep:`0048`)
 ##########################################################
 
 This module provides support for storing and retrieving bookmarks on
@@ -37,50 +38,36 @@ Service
 XSOs
 ====
 
-The following XSOs are used to represent an manipulate bookmark lists.
+All bookmark types must adhere to the following ABC:
 
-.. autoclass:: Storage
+.. autoclass:: Bookmark
+
+The following XSOs are used to represent an manipulate bookmark lists.
 
 .. autoclass:: Conference
 
 .. autoclass:: URL
 
+The following is used internally as the XSO container for bookmarks. If
+you want to support custom bookmark types, they have to be registered
+as children with the :attr:`Storage.bookmarks` descriptor:
+
+.. autoclass:: Storage
+
 Notes on usage
 ==============
 
-It is important to use this class carefully to prevent race conditions
-with other clients modifying the bookmark data (unfortunately, this is
-not entirely preventable due to the design of the bookmark protocol).
+.. currentmodule:: aioxmpp
 
-The recommended practice is to modify the bookmarks in a get, modify,
-set fashion, where the modify step should be as short as possible
-(that is, should not depend on user input).
+It is highly recommended to interact with the bookmark client via the
+provided signals and the get-modify-set methods
+:meth:`~BookmarkClient.add_bookmark`,
+:meth:`~BookmarkClient.remove_bookmark` and
+:meth:`~BookmarkClient.update_bookmark`.  Using
+:meth:`~BookmarkClient.set_bookmarks` directly is error prone and
+might cause data loss due to race conditions.
 
-The individual bookmark classes support comparison by value. This
-allows useful manipulation of the bookmark list in agreement with
-the get-modify-set pattern. Where we reference the objects to operate
-on by with the value retrieved in an earlier get.
-
-Removal::
-
-  storage = bookmark_client.get_bookmarks()
-  storage.bookmarks.remove(old_bookmark)
-  bookmark_client.set_bookmarks(storage)
-
-Modifying::
-
-  storage = bookmark_client.get_bookmarks()
-  i = storage.bookmarks.index(old_bookmark)
-  storage.bookmarks[i].name = "New Shiny Name"
-  storage.bookmarks[i].nick = "new_nick"
-  bookmark_client.set_bookmarks(storage)
-
-Adding::
-
-  storage = bookmark_client.get_bookmarks()
-  storage.bookmarks.append(new_bookmark)
-  bookmark_client.set_bookmarks(storage)
 """
 
-from .xso import (Storage, Conference, URL)  # NOQA
+from .xso import (Storage, Bookmark, Conference, URL)  # NOQA
 from .service import BookmarkClient  # NOQA