+++ aioxmpp/bookmarks/service.py
@@ -21,11 +21,8 @@
 ########################################################################
 import asyncio
 
-import aioxmpp
 import aioxmpp.callbacks as callbacks
 import aioxmpp.service as service
-import aioxmpp.disco as disco
-import aioxmpp.pubsub as pubsub
 import aioxmpp.private_xml as private_xml
 
 from . import xso as bookmark_xso
@@ -33,19 +30,65 @@ from . import xso as bookmark_xso
 
 # TODO: use private storage in pubsub where available.
 # TODO: sync bookmarks between pubsub and private xml storage
+# TODO: do we need merge-capabilities to reconcile the bookmarks
+# from different sources (local bookmark storage, pubsub, private xml
+# storage)
 class BookmarkClient(service.Service):
     """
     Supports retrieval and storage of bookmarks on the server.
     It currently only supports :xep:`Private XML Storage <49>` as
     backend.
 
+    There is the general rule *never* to modify the bookmark instances
+    retrieved from this class (either by :meth:`get_bookmarks` or as
+    an argument to one of the signals). If you need to modify a bookmark
+    for use with :meth:`update_bookmark` use :func:`copy.copy` to create
+    a copy.
+
+    .. automethod:: sync
+
     .. automethod:: get_bookmarks
 
     .. automethod:: set_bookmarks
 
+    The following methods change the bookmark list in a get-modify-set
+    pattern, to mitigate the danger of race conditions and should be
+    used in most circumstances:
+
+    .. automethod:: add_bookmark
+
+    .. automethod:: remove_bookmark
+
+    .. automethod:: update_bookmark
+
+
+    The following signals are provided that allow tracking the changes to
+    the bookmark list:
+
+    .. signal:: on_bookmark_added(added_bookmark)
+
+        Fires when a new bookmark is added.
+
+    .. signal:: on_bookmark_removed(removed_bookmark)
+
+        Fires when a bookmark is removed.
+
+    .. signal:: on_bookmark_changed(old_bookmark, new_bookmark)
+
+        Fires when a bookmark is changed.
+
+    .. note:: A heuristic is used to determine the change of bookmarks
+              and the reported changes may not directly reflect the
+              used methods, but it will always be possible to
+              construct the list of bookmarks from the events. For
+              example, when using :meth:`update_bookmark` to change
+              the JID of a :class:`Conference` bookmark a removed and
+              a added signal will fire.
+
     .. note:: The bookmark protocol is prone to race conditions if
               several clients access it concurrently. Be careful to
-              use a get-modify-set pattern.
+              use a get-modify-set pattern or the provided highlevel
+              interface.
 
     .. note:: Some other clients extend the bookmark format. For now
               those extensions are silently dropped by our XSOs, and
@@ -71,7 +114,7 @@ class BookmarkClient(service.Service):
         """
         Get the stored bookmarks from the server.
 
-        :returns: the bookmarks as a :class:`~bookmark_xso.Storage` object
+        :returns: a list of bookmarks
         """
         res = yield from self._private_xml.get_private_xml(
             bookmark_xso.Storage()
@@ -94,10 +137,15 @@ class BookmarkClient(service.Service):
         needed and set the bookmark cache to the new data.
         """
 
-        self.logger.debug("diffing %s, %s", self._bookmark_cache, new_bookmarks)
+        self.logger.debug("diffing %s, %s", self._bookmark_cache,
+                          new_bookmarks)
 
         def subdivide(level, old, new):
-
+            """
+            Subdivide the bookmarks according to the data item
+            ``bookmark.secondary[level]`` and emit the appropriate
+            events.
+            """
             if len(old) == len(new) == 1:
                 old_entry = old.pop()
                 new_entry = new.pop()
@@ -212,9 +260,10 @@ class BookmarkClient(service.Service):
     @asyncio.coroutine
     def get_bookmarks(self):
         """
-        Get the stored bookmarks from the server and emit signals.
+        Get the stored bookmarks from the server. Causes signals to be
+        fired to reflect the changes.
 
-        :returns: the bookmarks as a :class:`~bookmark_xso.Storage` object
+        :returns: a list of bookmarks
         """
         bookmarks = yield from self._get_bookmarks()
         self._diff_emit_update(bookmarks)
@@ -223,7 +272,18 @@ class BookmarkClient(service.Service):
     @asyncio.coroutine
     def set_bookmarks(self, bookmarks):
         """
-        Set the stored bookmarks.
+        Store the sequence of bookmarks `bookmarks`.
+
+        Causes signals to be fired to reflect the changes.
+
+        .. note:: This should normally not be used. It does not
+                  mitigate the race condition between clients
+                  concurrently modifying the bookmarks and may lead to
+                  data loss. Use :meth:`add_bookmark`,
+                  :meth:`remove_bookmark` and :meth:`update_bookmark`
+                  instead. This method still has use-cases (modifying
+                  the bookmarklist at large, e.g. by syncing the
+                  remote store with local data).
         """
         yield from self._set_bookmarks(bookmarks)
         self._diff_emit_update(bookmarks)
@@ -235,8 +295,8 @@ class BookmarkClient(service.Service):
         Sync the bookmarks between the local representation and the
         server.
 
-        This must be called periodically to assure that change events
-        are emitted.
+        This must be called periodically to assure that the signals
+        are fired.
         """
         yield from self.get_bookmarks()
 
@@ -283,8 +343,8 @@ class BookmarkClient(service.Service):
         Update a bookmark.
 
         The bookmark matches an existing bookmark `old` according to
-        bookmark equalitiy and replaces it by `new`. The bookmark is
-        added if no bookmark matching `old` exists.
+        bookmark equalitiy and replaces it by `new`. The bookmark
+        `new` is added if no bookmark matching `old` exists.
         """
         bookmarks = yield from self._get_bookmarks()
         for i, bookmark in enumerate(bookmarks):